Refine RUM injection offset writes
